+++ langchain_postgres/vectorstores.py
@@ -1,5 +1,6 @@
 from __future__ import annotations
 
+import asyncio
 import enum
 import logging
 import uuid
@@ -11,14 +12,15 @@ from typing import (
     List,
     Optional,
     Tuple,
-    Type,
-    Union,
+    Type, Union,
 )
 
 import numpy as np
 import sqlalchemy
-from sqlalchemy import SQLColumnExpression, cast, delete, func
+from sqlalchemy import SQLColumnExpression, cast, delete, func, select, Engine
 from sqlalchemy.dialects.postgresql import JSON, JSONB, JSONPATH, UUID, insert
+# TODO: accepter l'absence de l'option async lors des imports
+from sqlalchemy.ext.asyncio import AsyncSession, async_sessionmaker
 from sqlalchemy.orm import Session, relationship, sessionmaker
 
 try:
@@ -28,7 +30,6 @@ except ImportError:
 
 from langchain_core.documents import Document
 from langchain_core.embeddings import Embeddings
-from langchain_core.runnables.config import run_in_executor
 from langchain_core.utils import get_from_dict_or_env
 from langchain_core.vectorstores import VectorStore
 
@@ -47,9 +48,7 @@ DEFAULT_DISTANCE_STRATEGY = DistanceStrategy.COSINE
 
 Base = declarative_base()  # type: Any
 
-
 _LANGCHAIN_DEFAULT_COLLECTION_NAME = "langchain"
-
 
 _classes: Any = None
 
@@ -110,16 +109,28 @@ def _get_embedding_collection_store(vector_dimension: Optional[int] = None) -> A
 
         @classmethod
         def get_by_name(
-            cls, session: Session, name: str
+                cls, session: Session, name: str
         ) -> Optional["CollectionStore"]:
             return session.query(cls).filter(cls.name == name).first()  # type: ignore
 
         @classmethod
+        async def aget_by_name(
+                cls, session: AsyncSession, name: str
+        ) -> Optional["CollectionStore"]:
+            stmt = select(cls).filter(cls.name == name)
+            # return await session.execute(stmt)  # FIXME
+            return (await session.execute(stmt)).scalars().first()  # FIXME
+            # stmt = select(cls).filter(cls.name == name)
+            # result = await session.execute(stmt)
+            # x = result.scalars()
+            # return session.query(cls).filter(cls.name == name).first()
+
+        @classmethod
         def get_or_create(
-            cls,
-            session: Session,
-            name: str,
-            cmetadata: Optional[dict] = None,
+                cls,
+                session: Session,
+                name: str,
+                cmetadata: Optional[dict] = None,
         ) -> Tuple["CollectionStore", bool]:
             """Get or create a collection.
             Returns:
@@ -132,7 +143,29 @@ def _get_embedding_collection_store(vector_dimension: Optional[int] = None) -> A
 
             collection = cls(name=name, cmetadata=cmetadata)
             session.add(collection)
-            session.commit()
+            session.commit()  # FIXME PPR semble utile
+            created = True
+            return collection, created
+
+        @classmethod
+        async def aget_or_create(
+                cls,
+                session: AsyncSession,
+                name: str,
+                cmetadata: Optional[dict] = None,
+        ) -> Tuple["CollectionStore", bool]:
+            """
+            Get or create a collection.
+            Returns [Collection, bool] where the bool is True if the collection was created.
+            """  # noqa: E501
+            created = False
+            collection = await cls.aget_by_name(session, name)
+            if collection:
+                return collection, created
+
+            collection = cls(name=name, cmetadata=cmetadata)
+            session.add(collection)
+            await session.commit()
             created = True
             return collection, created
 
@@ -235,20 +268,21 @@ class PGVector(VectorStore):
     """
 
     def __init__(
-        self,
-        embeddings: Embeddings,
-        *,
-        connection: Optional[Connection] = None,
-        embedding_length: Optional[int] = None,
-        collection_name: str = _LANGCHAIN_DEFAULT_COLLECTION_NAME,
-        collection_metadata: Optional[dict] = None,
-        distance_strategy: DistanceStrategy = DEFAULT_DISTANCE_STRATEGY,
-        pre_delete_collection: bool = False,
-        logger: Optional[logging.Logger] = None,
-        relevance_score_fn: Optional[Callable[[float], float]] = None,
-        engine_args: Optional[dict[str, Any]] = None,
-        use_jsonb: bool = True,
-        create_extension: bool = True,
+            self,
+            embeddings: Embeddings,
+            *,
+            connection: Union[None, Connection, str] = None,
+            embedding_length: Optional[int] = None,
+            collection_name: str = _LANGCHAIN_DEFAULT_COLLECTION_NAME,
+            collection_metadata: Optional[dict] = None,
+            distance_strategy: DistanceStrategy = DEFAULT_DISTANCE_STRATEGY,
+            pre_delete_collection: bool = False,
+            logger: Optional[logging.Logger] = None,
+            relevance_score_fn: Optional[Callable[[float], float]] = None,
+            engine_args: Optional[dict[str, Any]] = None,
+            use_jsonb: bool = True,
+            create_extension: bool = True,
+            async_mode: bool = False,  # FIXME: a virer. Gaff aux imports sans async
     ) -> None:
         """Initialize the PGVector store.
 
@@ -277,6 +311,7 @@ class PGVector(VectorStore):
                 doesn't exist. disabling creation is useful when using ReadOnly
                 Databases.
         """
+        self.async_mode = async_mode
         self.embedding_function = embeddings
         self._embedding_length = embedding_length
         self.collection_name = collection_name
@@ -287,9 +322,8 @@ class PGVector(VectorStore):
         self.override_relevance_score_fn = relevance_score_fn
 
         if isinstance(connection, str):
-            self._engine = sqlalchemy.create_engine(
-                url=connection, **(engine_args or {})
-            )
+            self._engine = self._create_engine(
+                connection, engine_args, async_mode)
         elif isinstance(connection, sqlalchemy.engine.Engine):
             self._engine = connection
         else:
@@ -297,8 +331,13 @@ class PGVector(VectorStore):
                 "connection should be a connection string or an instance of "
                 "sqlalchemy.engine.Engine"
             )
-
-        self._session_maker = sessionmaker(bind=self._engine)
+        # If the driver accept only the synchrone calls, update the async_mode
+        self.async_mode = not isinstance(self._engine, Engine)
+        self._session_maker: Union[sessionmaker, async_sessionmaker]
+        if self.async_mode:
+            self._session_maker = async_sessionmaker(bind=self._engine)
+        else:
+            self._session_maker = sessionmaker(bind=self._engine)
 
         self.use_jsonb = use_jsonb
         self.create_extension = create_extension
@@ -306,10 +345,11 @@ class PGVector(VectorStore):
         if not use_jsonb:
             # Replace with a deprecation warning.
             raise NotImplementedError("use_jsonb=False is no longer supported.")
-        self.__post_init__()
+        if not async_mode:
+            self.__post_init__()
 
     def __post_init__(
-        self,
+            self,
     ) -> None:
         """Initialize the store."""
         if self.create_extension:
@@ -323,9 +363,85 @@ class PGVector(VectorStore):
         self.create_tables_if_not_exists()
         self.create_collection()
 
+    async def __apost_init__(
+            self,
+    ) -> None:
+
+        """Initialize the store."""
+        if self.create_extension:
+            await self.acreate_vector_extension()
+
+        EmbeddingStore, CollectionStore = _get_embedding_collection_store(
+            self._embedding_length
+        )
+        self.CollectionStore = CollectionStore
+        self.EmbeddingStore = EmbeddingStore
+        await self.acreate_tables_if_not_exists()
+        await self.acreate_collection()
+
+    @classmethod
+    async def create(cls,
+                     embeddings: Embeddings,
+                     *,
+                     connection: Optional[Connection] = None,
+                     embedding_length: Optional[int] = None,
+                     collection_name: str = _LANGCHAIN_DEFAULT_COLLECTION_NAME,
+                     collection_metadata: Optional[dict] = None,
+                     distance_strategy: DistanceStrategy = DEFAULT_DISTANCE_STRATEGY,
+                     pre_delete_collection: bool = False,
+                     logger: Optional[logging.Logger] = None,
+                     relevance_score_fn: Optional[Callable[[float], float]] = None,
+                     engine_args: Optional[dict[str, Any]] = None,
+                     use_jsonb: bool = True,
+                     create_extension: bool = True,
+                     async_mode: bool = True,
+                     ) -> PGVector:
+        self = cls(
+            embeddings=embeddings,
+            connection=connection,
+            embedding_length=embedding_length,
+            collection_name=collection_name,
+            collection_metadata=collection_metadata,
+            distance_strategy=distance_strategy,
+            pre_delete_collection=pre_delete_collection,
+            logger=logger,
+            relevance_score_fn=relevance_score_fn,
+            engine_args=engine_args,
+            use_jsonb=use_jsonb,
+            create_extension=create_extension,
+            async_mode=async_mode,
+        )
+        if async_mode:
+            await self.__apost_init__()
+        return self
+
+    def _create_engine(self,
+                       connection: str,
+                       engine_args: Optional[dict[str, Any]] = None,
+                       async_mode: bool = False) -> sqlalchemy.engine.Engine | sqlalchemy.ext.asyncio.AsyncEngine:
+        if async_mode:
+            from sqlalchemy.ext.asyncio import create_async_engine
+            # FIXME: gérer appel async sur un sync
+            from sqlalchemy.exc import InvalidRequestError
+            try:
+                return create_async_engine(
+                    url=connection,
+                    isolation_level="REPEATABLE READ",  # FIXME: merge avec la suite ?
+                    echo=True,  # FIXME: a virer
+                    **(engine_args or {})
+                )
+            except InvalidRequestError:
+                pass  # Ignore and return the synchrone version
+                logging.warning("Use a synchrone SQL engine !")
+        return sqlalchemy.create_engine(url=connection,
+                                        **(engine_args or {}))
+
     def __del__(self) -> None:
         if isinstance(self._engine, sqlalchemy.engine.Connection):
-            self._engine.close()
+            if self.async_mode:
+                asyncio.run(self._engine.close())
+            else:
+                self._engine.close()
 
     @property
     def embeddings(self) -> Embeddings:
@@ -350,22 +466,79 @@ class PGVector(VectorStore):
         except Exception as e:
             raise Exception(f"Failed to create vector extension: {e}") from e
 
+    async def acreate_vector_extension(self) -> None:
+        try:
+            async with self._session_maker() as session:
+                # The advisor lock fixes issue arising from concurrent
+                # creation of the vector extension.
+                # https://github.com/langchain-ai/langchain/issues/12933
+                # For more information see:
+                # https://www.postgresql.org/docs/16/explicit-locking.html#ADVISORY-LOCKS
+                await session.execute(
+                    sqlalchemy.text(
+                        "SELECT pg_advisory_xact_lock(1573678846307946496)"))
+                await session.execute(
+                    sqlalchemy.text("CREATE EXTENSION IF NOT EXISTS vector"))
+        except Exception as e:
+            raise Exception(f"Failed to create vector extension: {e}") from e
+
     def create_tables_if_not_exists(self) -> None:
         with self._session_maker() as session:
             Base.metadata.create_all(session.get_bind())
+
+    async def acreate_tables_if_not_exists(self) -> None:
+        if isinstance(self._engine, sqlalchemy.ext.asyncio.engine.AsyncConnection):
+            await self._engine.run_sync(Base.metadata.create_all)
+        else:
+            async with self._engine.begin() as conn:  # FIXME: session.run_sync existe
+                await conn.run_sync(Base.metadata.create_all)
+        # async with self._amake_session() as session:
+        #     await session.run_sync(Base.metadata.create_all)
 
     def drop_tables(self) -> None:
         with self._session_maker() as session:
             Base.metadata.drop_all(session.get_bind())
 
+    async def adrop_tables(self) -> None:
+        async with self._engine.begin() as conn:
+            await conn.run_sync(Base.metadata.drop_all)
+
     def create_collection(self) -> None:
         if self.pre_delete_collection:
             self.delete_collection()
-        with self._session_maker() as session:  # type: ignore[arg-type]
+        with self._session_maker() as session:
             self.CollectionStore.get_or_create(
                 session, self.collection_name, cmetadata=self.collection_metadata
             )
 
+    async def acreate_collection(self) -> None:
+        async with self._session_maker() as session:
+            if self.pre_delete_collection:
+                await self._adelete_collection(session)
+            await self.CollectionStore.aget_or_create(
+                session, self.collection_name, cmetadata=self.collection_metadata
+            )
+
+    def _delete_collection(self, session: Session) -> None:
+        self.logger.debug("Trying to delete collection")
+        collection = self.get_collection(session)
+        if not collection:
+            self.logger.warning("Collection not found")
+            return
+        session.delete(collection)
+
+    # FIXME: necessaire le _adelete ?
+    async def _adelete_collection(self, session: AsyncSession) -> None:
+        self.logger.debug("Trying to delete collection")
+        collection = await self.aget_collection(session)
+        if not collection:
+            self.logger.warning("Collection not found")
+            return
+        await session.delete(collection)
+
+    # def delete_collection(self) -> None:
+    #     with self._session_maker() as session:
+    #         self._delete_collection(session)
     def delete_collection(self) -> None:
         self.logger.debug("Trying to delete collection")
         with self._session_maker() as session:  # type: ignore[arg-type]
@@ -376,11 +549,21 @@ class PGVector(VectorStore):
             session.delete(collection)
             session.commit()
 
+    async def adelete_collection(self) -> None:
+        self.logger.debug("Trying to delete collection")
+        with self._session_maker() as session:  # type: ignore[arg-type]
+            collection = await self.aget_collection(session)
+            if not collection:
+                self.logger.warning("Collection not found")
+                return
+            await session.adelete(collection)
+            await session.commit()
+
     def delete(
-        self,
-        ids: Optional[List[str]] = None,
-        collection_only: bool = False,
-        **kwargs: Any,
+            self,
+            ids: Optional[List[str]] = None,
+            collection_only: bool = False,
+            **kwargs: Any,
     ) -> None:
         """Delete vectors by ids or uuids.
 
@@ -411,24 +594,62 @@ class PGVector(VectorStore):
                 session.execute(stmt)
             session.commit()
 
+    async def adelete(
+            self,
+            ids: Optional[List[str]] = None,
+            collection_only: bool = False,
+            **kwargs: Any,
+    ) -> None:
+        """Delete vectors by ids or uuids.
+
+        Args:
+            ids: List of ids to delete.
+            collection_only: Only delete ids in the collection.
+        """
+        async with self._session_maker() as session:
+            if ids is not None:
+                self.logger.debug(
+                    "Trying to delete vectors by ids (represented by the model "
+                    "using the custom ids field)"
+                )
+
+                stmt = delete(self.EmbeddingStore)
+
+                if collection_only:
+                    collection = await self.aget_collection(session)
+                    if not collection:
+                        self.logger.warning("Collection not found")
+                        return
+
+                    stmt = stmt.where(
+                        self.EmbeddingStore.collection_id == collection.uuid
+                    )
+
+                stmt = stmt.where(self.EmbeddingStore.id.in_(ids))
+                await session.execute(stmt)
+            await session.commit()
+
     def get_collection(self, session: Session) -> Any:
         return self.CollectionStore.get_by_name(session, self.collection_name)
 
+    async def aget_collection(self, session: AsyncSession) -> Any:
+        return await self.CollectionStore.aget_by_name(session, self.collection_name)
+
     @classmethod
     def __from(
-        cls,
-        texts: List[str],
-        embeddings: List[List[float]],
-        embedding: Embeddings,
-        metadatas: Optional[List[dict]] = None,
-        ids: Optional[List[str]] = None,
-        collection_name: str = _LANGCHAIN_DEFAULT_COLLECTION_NAME,
-        distance_strategy: DistanceStrategy = DEFAULT_DISTANCE_STRATEGY,
-        connection: Optional[str] = None,
-        pre_delete_collection: bool = False,
-        *,
-        use_jsonb: bool = True,
-        **kwargs: Any,
+            cls,
+            texts: List[str],
+            embeddings: List[List[float]],
+            embedding: Embeddings,
+            metadatas: Optional[List[dict]] = None,
+            ids: Optional[List[str]] = None,
+            collection_name: str = _LANGCHAIN_DEFAULT_COLLECTION_NAME,
+            distance_strategy: DistanceStrategy = DEFAULT_DISTANCE_STRATEGY,
+            connection: Optional[str] = None,
+            pre_delete_collection: bool = False,
+            *,
+            use_jsonb: bool = True,
+            **kwargs: Any,
     ) -> PGVector:
         if ids is None:
             ids = [str(uuid.uuid4()) for _ in texts]
@@ -452,13 +673,54 @@ class PGVector(VectorStore):
 
         return store
 
+    @classmethod
+    async def __afrom(
+            cls,
+            texts: List[str],
+            embeddings: List[List[float]],
+            embedding: Embeddings,
+            metadatas: Optional[List[dict]] = None,
+            ids: Optional[List[str]] = None,
+            collection_name: str = _LANGCHAIN_DEFAULT_COLLECTION_NAME,
+            distance_strategy: DistanceStrategy = DEFAULT_DISTANCE_STRATEGY,
+            connection: Optional[str] = None,
+            pre_delete_collection: bool = False,
+            *,
+            use_jsonb: bool = True,
+            **kwargs: Any,
+    ) -> PGVector:
+        if ids is None:
+            ids = [str(uuid.uuid1()) for _ in texts]
+
+        if not metadatas:
+            metadatas = [{} for _ in texts]
+
+        store = cls(
+            connection=connection,
+            collection_name=collection_name,
+            embeddings=embedding,
+            distance_strategy=distance_strategy,
+            pre_delete_collection=pre_delete_collection,
+            use_jsonb=use_jsonb,
+            async_mode=True,  # FIXME
+            **kwargs,
+        )
+        # Second phase to create
+        await store.__apost_init__()
+
+        await store.aadd_embeddings(
+            texts=texts, embeddings=embeddings, metadatas=metadatas, ids=ids, **kwargs
+        )
+
+        return store
+
     def add_embeddings(
-        self,
-        texts: Iterable[str],
-        embeddings: List[List[float]],
-        metadatas: Optional[List[dict]] = None,
-        ids: Optional[List[str]] = None,
-        **kwargs: Any,
+            self,
+            texts: Iterable[str],
+            embeddings: List[List[float]],
+            metadatas: Optional[List[dict]] = None,
+            ids: Optional[List[str]] = None,
+            **kwargs: Any,
     ) -> List[str]:
         """Add embeddings to the vectorstore.
 
@@ -505,12 +767,65 @@ class PGVector(VectorStore):
 
         return ids
 
+    async def aadd_embeddings(
+            self,
+            texts: Iterable[str],
+            embeddings: List[List[float]],
+            metadatas: Optional[List[dict]] = None,
+            ids: Optional[List[str]] = None,
+            **kwargs: Any,
+    ) -> List[str]:
+        """Add embeddings to the vectorstore.
+
+        Args:
+            texts: Iterable of strings to add to the vectorstore.
+            embeddings: List of list of embedding vectors.
+            metadatas: List of metadatas associated with the texts.
+            kwargs: vectorstore specific parameters
+        """
+        if ids is None:
+            ids = [str(uuid.uuid1()) for _ in texts]
+
+        if not metadatas:
+            metadatas = [{} for _ in texts]
+
+        async with self._session_maker() as session:  # type: ignore[arg-type]
+            collection = await self.aget_collection(session)
+            if not collection:
+                raise ValueError("Collection not found")
+            data = [
+                {
+                    "id": id,
+                    "collection_id": collection.uuid,
+                    "embedding": embedding,
+                    "document": text,
+                    "cmetadata": metadata or {},
+                }
+                for text, metadata, embedding, id in zip(
+                    texts, metadatas, embeddings, ids
+                )
+            ]
+            stmt = insert(self.EmbeddingStore).values(data)
+            on_conflict_stmt = stmt.on_conflict_do_update(
+                index_elements=["id"],
+                # Conflict detection based on these columns
+                set_={
+                    "embedding": stmt.excluded.embedding,
+                    "document": stmt.excluded.document,
+                    "cmetadata": stmt.excluded.cmetadata,
+                },
+            )
+            await session.execute(on_conflict_stmt)
+            await session.commit()
+
+        return ids
+
     def add_texts(
-        self,
-        texts: Iterable[str],
-        metadatas: Optional[List[dict]] = None,
-        ids: Optional[List[str]] = None,
-        **kwargs: Any,
+            self,
+            texts: Iterable[str],
+            metadatas: Optional[List[dict]] = None,
+            ids: Optional[List[str]] = None,
+            **kwargs: Any,
     ) -> List[str]:
         """Run more texts through the embeddings and add to the vectorstore.
 
@@ -527,12 +842,34 @@ class PGVector(VectorStore):
             texts=texts, embeddings=embeddings, metadatas=metadatas, ids=ids, **kwargs
         )
 
+    async def aadd_texts(
+            self,
+            texts: Iterable[str],
+            metadatas: Optional[List[dict]] = None,
+            ids: Optional[List[str]] = None,
+            **kwargs: Any,
+    ) -> List[str]:
+        """Run more texts through the embeddings and add to the vectorstore.
+
+        Args:
+            texts: Iterable of strings to add to the vectorstore.
+            metadatas: Optional list of metadatas associated with the texts.
+            kwargs: vectorstore specific parameters
+
+        Returns:
+            List of ids from adding the texts into the vectorstore.
+        """
+        embeddings = await self.embedding_function.aembed_documents(list(texts))
+        return await self.aadd_embeddings(
+            texts=texts, embeddings=embeddings, metadatas=metadatas, ids=ids, **kwargs
+        )
+
     def similarity_search(
-        self,
-        query: str,
-        k: int = 4,
-        filter: Optional[dict] = None,
-        **kwargs: Any,
+            self,
+            query: str,
+            k: int = 4,
+            filter: Optional[dict] = None,
+            **kwargs: Any,
     ) -> List[Document]:
         """Run similarity search with PGVector with distance.
 
@@ -551,11 +888,35 @@ class PGVector(VectorStore):
             filter=filter,
         )
 
+    async def asimilarity_search(
+            self,
+            query: str,
+            k: int = 4,
+            filter: Optional[dict] = None,
+            **kwargs: Any,
+    ) -> List[Document]:
+        """Run similarity search with PGVector with distance.
+
+        Args:
+            query (str): Query text to search for.
+            k (int): Number of results to return. Defaults to 4.
+            filter (Optional[Dict[str, str]]): Filter by metadata. Defaults to None.
+
+        Returns:
+            List of Documents most similar to the query.
+        """
+        embedding = self.embedding_function.embed_query(text=query)
+        return await self.asimilarity_search_by_vector(
+            embedding=embedding,
+            k=k,
+            filter=filter,
+        )
+
     def similarity_search_with_score(
-        self,
-        query: str,
-        k: int = 4,
-        filter: Optional[dict] = None,
+            self,
+            query: str,
+            k: int = 4,
+            filter: Optional[dict] = None,
     ) -> List[Tuple[Document, float]]:
         """Return docs most similar to query.
 
@@ -569,6 +930,28 @@ class PGVector(VectorStore):
         """
         embedding = self.embedding_function.embed_query(query)
         docs = self.similarity_search_with_score_by_vector(
+            embedding=embedding, k=k, filter=filter
+        )
+        return docs
+
+    async def asimilarity_search_with_score(
+            self,
+            query: str,
+            k: int = 4,
+            filter: Optional[dict] = None,
+    ) -> List[Tuple[Document, float]]:
+        """Return docs most similar to query.
+
+        Args:
+            query: Text to look up documents similar to.
+            k: Number of Documents to return. Defaults to 4.
+            filter (Optional[Dict[str, str]]): Filter by metadata. Defaults to None.
+
+        Returns:
+            List of Documents most similar to the query and score for each.
+        """
+        embedding = self.embedding_function.embed_query(query)
+        docs = await self.asimilarity_search_with_score_by_vector(
             embedding=embedding, k=k, filter=filter
         )
         return docs
@@ -588,14 +971,28 @@ class PGVector(VectorStore):
             )
 
     def similarity_search_with_score_by_vector(
-        self,
-        embedding: List[float],
-        k: int = 4,
-        filter: Optional[dict] = None,
+            self,
+            embedding: List[float],
+            k: int = 4,
+            filter: Optional[dict] = None,
     ) -> List[Tuple[Document, float]]:
         results = self.__query_collection(embedding=embedding, k=k, filter=filter)
 
         return self._results_to_docs_and_scores(results)
+
+    async def asimilarity_search_with_score_by_vector(
+            self,
+            embedding: List[float],
+            k: int = 4,
+            filter: Optional[dict] = None,
+    ) -> List[Tuple[Document, float]]:
+        async with self._session_maker() as session:  # type: ignore[arg-type]
+            results = await self.__aquery_collection(
+                session=session,
+                embedding=embedding, k=k,
+                filter=filter)
+
+            return self._results_to_docs_and_scores(results)
 
     def _results_to_docs_and_scores(self, results: Any) -> List[Tuple[Document, float]]:
         """Return docs and scores from results."""
@@ -612,9 +1009,9 @@ class PGVector(VectorStore):
         return docs
 
     def _handle_field_filter(
-        self,
-        field: str,
-        value: Any,
+            self,
+            field: str,
+            value: Any,
     ) -> SQLColumnExpression:
         """Create a filter for a specific field.
 
@@ -724,7 +1121,8 @@ class PGVector(VectorStore):
         else:
             raise NotImplementedError()
 
-    def _create_filter_clause_deprecated(self, key, value):  # type: ignore[no-untyped-def]
+    def _create_filter_clause_deprecated(self, key,
+                                         value):  # type: ignore[no-untyped-def]
         """Deprecated functionality.
 
         This is for backwards compatibility with the JSON based schema for metadata.
@@ -793,7 +1191,7 @@ class PGVector(VectorStore):
         return filter_by_metadata
 
     def _create_filter_clause_json_deprecated(
-        self, filter: Any
+            self, filter: Any
     ) -> List[SQLColumnExpression]:
         """Convert filters from IR to SQL clauses.
 
@@ -904,10 +1302,10 @@ class PGVector(VectorStore):
             )
 
     def __query_collection(
-        self,
-        embedding: List[float],
-        k: int = 4,
-        filter: Optional[Dict[str, str]] = None,
+            self,
+            embedding: List[float],
+            k: int = 4,
+            filter: Optional[Dict[str, str]] = None,
     ) -> List[Any]:
         """Query the collection."""
         with self._session_maker() as session:  # type: ignore[arg-type]
@@ -945,12 +1343,53 @@ class PGVector(VectorStore):
 
         return results
 
+    async def __aquery_collection(
+            self,
+            session: AsyncSession,
+            embedding: List[float],
+            k: int = 4,
+            filter: Optional[Dict[str, str]] = None,
+    ) -> List[Any]:
+        """Query the collection."""
+        collection = await self.aget_collection(session)
+        if not collection:
+            raise ValueError("Collection not found")
+
+        filter_by = [self.EmbeddingStore.collection_id == collection.uuid]
+        if filter:
+            if self.use_jsonb:
+                filter_clauses = self._create_filter_clause(filter)
+                if filter_clauses is not None:
+                    filter_by.append(filter_clauses)
+            else:
+                # Old way of doing things
+                filter_clauses = self._create_filter_clause_json_deprecated(filter)
+                filter_by.extend(filter_clauses)
+
+        _type = self.EmbeddingStore
+
+        stmt = (select(
+            self.EmbeddingStore,
+            self.distance_strategy(embedding).label("distance"),  # type: ignore
+        )
+                .filter(*filter_by)
+                .order_by(sqlalchemy.asc("distance"))
+                .join(
+            self.CollectionStore,
+            self.EmbeddingStore.collection_id == self.CollectionStore.uuid,
+        )
+                .limit(k))
+
+        results: List[Any] = (await session.execute(stmt)).all()
+
+        return results
+
     def similarity_search_by_vector(
-        self,
-        embedding: List[float],
-        k: int = 4,
-        filter: Optional[dict] = None,
-        **kwargs: Any,
+            self,
+            embedding: List[float],
+            k: int = 4,
+            filter: Optional[dict] = None,
+            **kwargs: Any,
     ) -> List[Document]:
         """Return docs most similar to embedding vector.
 
@@ -967,19 +1406,41 @@ class PGVector(VectorStore):
         )
         return _results_to_docs(docs_and_scores)
 
+    async def asimilarity_search_by_vector(
+            self,
+            embedding: List[float],
+            k: int = 4,
+            filter: Optional[dict] = None,
+            **kwargs: Any,
+    ) -> List[Document]:
+        """Return docs most similar to embedding vector.
+
+        Args:
+            embedding: Embedding to look up documents similar to.
+            k: Number of Documents to return. Defaults to 4.
+            filter (Optional[Dict[str, str]]): Filter by metadata. Defaults to None.
+
+        Returns:
+            List of Documents most similar to the query vector.
+        """
+        docs_and_scores = await self.asimilarity_search_with_score_by_vector(
+            embedding=embedding, k=k, filter=filter
+        )
+        return _results_to_docs(docs_and_scores)
+
     @classmethod
     def from_texts(
-        cls: Type[PGVector],
-        texts: List[str],
-        embedding: Embeddings,
-        metadatas: Optional[List[dict]] = None,
-        *,
-        collection_name: str = _LANGCHAIN_DEFAULT_COLLECTION_NAME,
-        distance_strategy: DistanceStrategy = DEFAULT_DISTANCE_STRATEGY,
-        ids: Optional[List[str]] = None,
-        pre_delete_collection: bool = False,
-        use_jsonb: bool = True,
-        **kwargs: Any,
+            cls: Type[PGVector],
+            texts: List[str],
+            embedding: Embeddings,
+            metadatas: Optional[List[dict]] = None,
+            *,
+            collection_name: str = _LANGCHAIN_DEFAULT_COLLECTION_NAME,
+            distance_strategy: DistanceStrategy = DEFAULT_DISTANCE_STRATEGY,
+            ids: Optional[List[str]] = None,
+            pre_delete_collection: bool = False,
+            use_jsonb: bool = True,
+            **kwargs: Any,
     ) -> PGVector:
         """Return VectorStore initialized from documents and embeddings."""
         embeddings = embedding.embed_documents(list(texts))
@@ -998,17 +1459,46 @@ class PGVector(VectorStore):
         )
 
     @classmethod
+    async def afrom_texts(
+            cls: Type[PGVector],
+            texts: List[str],
+            embedding: Embeddings,
+            metadatas: Optional[List[dict]] = None,
+            collection_name: str = _LANGCHAIN_DEFAULT_COLLECTION_NAME,
+            distance_strategy: DistanceStrategy = DEFAULT_DISTANCE_STRATEGY,
+            ids: Optional[List[str]] = None,
+            pre_delete_collection: bool = False,
+            *,
+            use_jsonb: bool = True,
+            **kwargs: Any,
+    ) -> PGVector:
+        """Return VectorStore initialized from documents and embeddings."""
+        embeddings = embedding.embed_documents(list(texts))
+        return await cls.__afrom(
+            texts,
+            embeddings,
+            embedding,
+            metadatas=metadatas,
+            ids=ids,
+            collection_name=collection_name,
+            distance_strategy=distance_strategy,
+            pre_delete_collection=pre_delete_collection,
+            use_jsonb=use_jsonb,
+            **kwargs,
+        )
+
+    @classmethod
     def from_embeddings(
-        cls,
-        text_embeddings: List[Tuple[str, List[float]]],
-        embedding: Embeddings,
-        *,
-        metadatas: Optional[List[dict]] = None,
-        collection_name: str = _LANGCHAIN_DEFAULT_COLLECTION_NAME,
-        distance_strategy: DistanceStrategy = DEFAULT_DISTANCE_STRATEGY,
-        ids: Optional[List[str]] = None,
-        pre_delete_collection: bool = False,
-        **kwargs: Any,
+            cls,
+            text_embeddings: List[Tuple[str, List[float]]],
+            embedding: Embeddings,
+            *,
+            metadatas: Optional[List[dict]] = None,
+            collection_name: str = _LANGCHAIN_DEFAULT_COLLECTION_NAME,
+            distance_strategy: DistanceStrategy = DEFAULT_DISTANCE_STRATEGY,
+            ids: Optional[List[str]] = None,
+            pre_delete_collection: bool = False,
+            **kwargs: Any,
     ) -> PGVector:
         """Construct PGVector wrapper from raw documents and embeddings.
 
@@ -1054,15 +1544,60 @@ class PGVector(VectorStore):
         )
 
     @classmethod
+    async def afrom_embeddings(
+            cls,
+            text_embeddings: List[Tuple[str, List[float]]],
+            embedding: Embeddings,
+            metadatas: Optional[List[dict]] = None,
+            collection_name: str = _LANGCHAIN_DEFAULT_COLLECTION_NAME,
+            distance_strategy: DistanceStrategy = DEFAULT_DISTANCE_STRATEGY,
+            ids: Optional[List[str]] = None,
+            pre_delete_collection: bool = False,
+            **kwargs: Any,
+    ) -> PGVector:
+        """Construct PGVector wrapper from raw documents and pre-
+        generated embeddings.
+
+        Return VectorStore initialized from documents and embeddings.
+        Postgres connection string is required
+        "Either pass it as a parameter
+        or set the PGVECTOR_CONNECTION_STRING environment variable.
+
+        Example:
+            .. code-block:: python
+
+                from langchain_community.vectorstores import PGVector
+                from langchain_community.embeddings import OpenAIEmbeddings
+                embeddings = OpenAIEmbeddings()
+                text_embeddings = embeddings.embed_documents(texts)
+                text_embedding_pairs = list(zip(texts, text_embeddings))
+                faiss = PGVector.from_embeddings(text_embedding_pairs, embeddings)
+        """
+        texts = [t[0] for t in text_embeddings]
+        embeddings = [t[1] for t in text_embeddings]
+
+        return await cls.__afrom(
+            texts,
+            embeddings,
+            embedding,
+            metadatas=metadatas,
+            ids=ids,
+            collection_name=collection_name,
+            distance_strategy=distance_strategy,
+            pre_delete_collection=pre_delete_collection,
+            **kwargs,
+        )
+
+    @classmethod
     def from_existing_index(
-        cls: Type[PGVector],
-        embedding: Embeddings,
-        *,
-        collection_name: str = _LANGCHAIN_DEFAULT_COLLECTION_NAME,
-        distance_strategy: DistanceStrategy = DEFAULT_DISTANCE_STRATEGY,
-        pre_delete_collection: bool = False,
-        connection: Optional[Connection] = None,
-        **kwargs: Any,
+            cls: Type[PGVector],
+            embedding: Embeddings,
+            *,
+            collection_name: str = _LANGCHAIN_DEFAULT_COLLECTION_NAME,
+            distance_strategy: DistanceStrategy = DEFAULT_DISTANCE_STRATEGY,
+            pre_delete_collection: bool = False,
+            connection: Optional[Connection] = None,
+            **kwargs: Any,
     ) -> PGVector:
         """
         Get instance of an existing PGVector store.This method will
@@ -1081,10 +1616,38 @@ class PGVector(VectorStore):
         return store
 
     @classmethod
+    async def afrom_existing_index(
+            cls: Type[PGVector],
+            embedding: Embeddings,
+            *,
+            collection_name: str = _LANGCHAIN_DEFAULT_COLLECTION_NAME,
+            distance_strategy: DistanceStrategy = DEFAULT_DISTANCE_STRATEGY,
+            pre_delete_collection: bool = False,
+            connection: Optional[Connection] = None,
+            **kwargs: Any,
+    ) -> PGVector:
+        """
+        Get instance of an existing PGVector store.This method will
+        return the instance of the store without inserting any new
+        embeddings
+        """
+        store = cls(  # FIXME: créate
+            connection=connection,
+            collection_name=collection_name,
+            embeddings=embedding,
+            distance_strategy=distance_strategy,
+            pre_delete_collection=pre_delete_collection,
+            async_mode=True,
+            **kwargs,
+        )
+
+        return store
+
+    @classmethod
     def get_connection_string(cls, kwargs: Dict[str, Any]) -> str:
         connection_string: str = get_from_dict_or_env(
             data=kwargs,
-            key="connection_string",
+            key="connection",
             env_key="PGVECTOR_CONNECTION_STRING",
         )
 
@@ -1099,17 +1662,17 @@ class PGVector(VectorStore):
 
     @classmethod
     def from_documents(
-        cls: Type[PGVector],
-        documents: List[Document],
-        embedding: Embeddings,
-        *,
-        connection: Optional[Connection] = None,
-        collection_name: str = _LANGCHAIN_DEFAULT_COLLECTION_NAME,
-        distance_strategy: DistanceStrategy = DEFAULT_DISTANCE_STRATEGY,
-        ids: Optional[List[str]] = None,
-        pre_delete_collection: bool = False,
-        use_jsonb: bool = True,
-        **kwargs: Any,
+            cls: Type[PGVector],
+            documents: List[Document],
+            embedding: Embeddings,
+            *,
+            connection: Optional[Connection] = None,
+            collection_name: str = _LANGCHAIN_DEFAULT_COLLECTION_NAME,
+            distance_strategy: DistanceStrategy = DEFAULT_DISTANCE_STRATEGY,
+            ids: Optional[List[str]] = None,
+            pre_delete_collection: bool = False,
+            use_jsonb: bool = True,
+            **kwargs: Any,
     ) -> PGVector:
         """Return VectorStore initialized from documents and embeddings."""
 
@@ -1130,14 +1693,52 @@ class PGVector(VectorStore):
         )
 
     @classmethod
+    async def afrom_documents(
+            cls: Type[PGVector],
+            documents: List[Document],
+            embedding: Embeddings,
+            collection_name: str = _LANGCHAIN_DEFAULT_COLLECTION_NAME,
+            distance_strategy: DistanceStrategy = DEFAULT_DISTANCE_STRATEGY,
+            ids: Optional[List[str]] = None,
+            pre_delete_collection: bool = False,
+            *,
+            use_jsonb: bool = True,
+            **kwargs: Any,
+    ) -> PGVector:
+        """
+        Return VectorStore initialized from documents and embeddings.
+        Postgres connection string is required
+        "Either pass it as a parameter
+        or set the PGVECTOR_CONNECTION_STRING environment variable.
+        """
+
+        texts = [d.page_content for d in documents]
+        metadatas = [d.metadata for d in documents]
+        connection_string = cls.get_connection_string(kwargs)
+
+        kwargs["connection"] = connection_string
+
+        return await cls.afrom_texts(
+            texts=texts,
+            pre_delete_collection=pre_delete_collection,
+            embedding=embedding,
+            distance_strategy=distance_strategy,
+            metadatas=metadatas,
+            ids=ids,
+            collection_name=collection_name,
+            use_jsonb=use_jsonb,
+            **kwargs,
+        )
+
+    @classmethod
     def connection_string_from_db_params(
-        cls,
-        driver: str,
-        host: str,
-        port: int,
-        database: str,
-        user: str,
-        password: str,
+            cls,
+            driver: str,
+            host: str,
+            port: int,
+            database: str,
+            user: str,
+            password: str,
     ) -> str:
         """Return connection string from database parameters."""
         if driver != "psycopg":
@@ -1172,13 +1773,13 @@ class PGVector(VectorStore):
             )
 
     def max_marginal_relevance_search_with_score_by_vector(
-        self,
-        embedding: List[float],
-        k: int = 4,
-        fetch_k: int = 20,
-        lambda_mult: float = 0.5,
-        filter: Optional[Dict[str, str]] = None,
-        **kwargs: Any,
+            self,
+            embedding: List[float],
+            k: int = 4,
+            fetch_k: int = 20,
+            lambda_mult: float = 0.5,
+            filter: Optional[Dict[str, str]] = None,
+            **kwargs: Any,
     ) -> List[Tuple[Document, float]]:
         """Return docs selected using the maximal marginal relevance with score
             to embedding vector.
@@ -1216,14 +1817,62 @@ class PGVector(VectorStore):
 
         return [r for i, r in enumerate(candidates) if i in mmr_selected]
 
+    async def amax_marginal_relevance_search_with_score_by_vector(
+            self,
+            embedding: List[float],
+            k: int = 4,
+            fetch_k: int = 20,
+            lambda_mult: float = 0.5,
+            filter: Optional[Dict[str, str]] = None,
+            **kwargs: Any,
+    ) -> List[Tuple[Document, float]]:
+        """Return docs selected using the maximal marginal relevance with score
+            to embedding vector.
+
+        Maximal marginal relevance optimizes for similarity to query AND diversity
+            among selected documents.
+
+        Args:
+            embedding: Embedding to look up documents similar to.
+            k (int): Number of Documents to return. Defaults to 4.
+            fetch_k (int): Number of Documents to fetch to pass to MMR algorithm.
+                Defaults to 20.
+            lambda_mult (float): Number between 0 and 1 that determines the degree
+                of diversity among the results with 0 corresponding
+                to maximum diversity and 1 to minimum diversity.
+                Defaults to 0.5.
+            filter (Optional[Dict[str, str]]): Filter by metadata. Defaults to None.
+
+        Returns:
+            List[Tuple[Document, float]]: List of Documents selected by maximal marginal
+                relevance to the query and score for each.
+        """
+        async with self._session_maker() as session:
+            results = await self.__aquery_collection(session=session,
+                                                     embedding=embedding, k=fetch_k,
+                                                     filter=filter)
+
+            embedding_list = [result.EmbeddingStore.embedding for result in results]
+
+            mmr_selected = maximal_marginal_relevance(
+                np.array(embedding, dtype=np.float32),
+                embedding_list,
+                k=k,
+                lambda_mult=lambda_mult,
+            )
+
+            candidates = self._results_to_docs_and_scores(results)
+
+            return [r for i, r in enumerate(candidates) if i in mmr_selected]
+
     def max_marginal_relevance_search(
-        self,
-        query: str,
-        k: int = 4,
-        fetch_k: int = 20,
-        lambda_mult: float = 0.5,
-        filter: Optional[Dict[str, str]] = None,
-        **kwargs: Any,
+            self,
+            query: str,
+            k: int = 4,
+            fetch_k: int = 20,
+            lambda_mult: float = 0.5,
+            filter: Optional[Dict[str, str]] = None,
+            **kwargs: Any,
     ) -> List[Document]:
         """Return docs selected using the maximal marginal relevance.
 
@@ -1254,14 +1903,52 @@ class PGVector(VectorStore):
             **kwargs,
         )
 
+    async def amax_marginal_relevance_search(
+            self,
+            query: str,
+            k: int = 4,
+            fetch_k: int = 20,
+            lambda_mult: float = 0.5,
+            filter: Optional[Dict[str, str]] = None,
+            **kwargs: Any,
+    ) -> List[Document]:
+        """Return docs selected using the maximal marginal relevance.
+
+        Maximal marginal relevance optimizes for similarity to query AND diversity
+            among selected documents.
+
+        Args:
+            query (str): Text to look up documents similar to.
+            k (int): Number of Documents to return. Defaults to 4.
+            fetch_k (int): Number of Documents to fetch to pass to MMR algorithm.
+                Defaults to 20.
+            lambda_mult (float): Number between 0 and 1 that determines the degree
+                of diversity among the results with 0 corresponding
+                to maximum diversity and 1 to minimum diversity.
+                Defaults to 0.5.
+            filter (Optional[Dict[str, str]]): Filter by metadata. Defaults to None.
+
+        Returns:
+            List[Document]: List of Documents selected by maximal marginal relevance.
+        """
+        embedding = self.embedding_function.embed_query(query)
+        return await self.amax_marginal_relevance_search_by_vector(
+            embedding,
+            k=k,
+            fetch_k=fetch_k,
+            lambda_mult=lambda_mult,
+            filter=filter,
+            **kwargs,
+        )
+
     def max_marginal_relevance_search_with_score(
-        self,
-        query: str,
-        k: int = 4,
-        fetch_k: int = 20,
-        lambda_mult: float = 0.5,
-        filter: Optional[dict] = None,
-        **kwargs: Any,
+            self,
+            query: str,
+            k: int = 4,
+            fetch_k: int = 20,
+            lambda_mult: float = 0.5,
+            filter: Optional[dict] = None,
+            **kwargs: Any,
     ) -> List[Tuple[Document, float]]:
         """Return docs selected using the maximal marginal relevance with score.
 
@@ -1294,14 +1981,54 @@ class PGVector(VectorStore):
         )
         return docs
 
+    async def amax_marginal_relevance_search_with_score(
+            self,
+            query: str,
+            k: int = 4,
+            fetch_k: int = 20,
+            lambda_mult: float = 0.5,
+            filter: Optional[dict] = None,
+            **kwargs: Any,
+    ) -> List[Tuple[Document, float]]:
+        """Return docs selected using the maximal marginal relevance with score.
+
+        Maximal marginal relevance optimizes for similarity to query AND diversity
+            among selected documents.
+
+        Args:
+            query (str): Text to look up documents similar to.
+            k (int): Number of Documents to return. Defaults to 4.
+            fetch_k (int): Number of Documents to fetch to pass to MMR algorithm.
+                Defaults to 20.
+            lambda_mult (float): Number between 0 and 1 that determines the degree
+                of diversity among the results with 0 corresponding
+                to maximum diversity and 1 to minimum diversity.
+                Defaults to 0.5.
+            filter (Optional[Dict[str, str]]): Filter by metadata. Defaults to None.
+
+        Returns:
+            List[Tuple[Document, float]]: List of Documents selected by maximal marginal
+                relevance to the query and score for each.
+        """
+        embedding = self.embedding_function.embed_query(query)
+        docs = await self.amax_marginal_relevance_search_with_score_by_vector(
+            embedding=embedding,
+            k=k,
+            fetch_k=fetch_k,
+            lambda_mult=lambda_mult,
+            filter=filter,
+            **kwargs,
+        )
+        return docs
+
     def max_marginal_relevance_search_by_vector(
-        self,
-        embedding: List[float],
-        k: int = 4,
-        fetch_k: int = 20,
-        lambda_mult: float = 0.5,
-        filter: Optional[Dict[str, str]] = None,
-        **kwargs: Any,
+            self,
+            embedding: List[float],
+            k: int = 4,
+            fetch_k: int = 20,
+            lambda_mult: float = 0.5,
+            filter: Optional[Dict[str, str]] = None,
+            **kwargs: Any,
     ) -> List[Document]:
         """Return docs selected using the maximal marginal relevance
             to embedding vector.
@@ -1335,22 +2062,35 @@ class PGVector(VectorStore):
         return _results_to_docs(docs_and_scores)
 
     async def amax_marginal_relevance_search_by_vector(
-        self,
-        embedding: List[float],
-        k: int = 4,
-        fetch_k: int = 20,
-        lambda_mult: float = 0.5,
-        filter: Optional[Dict[str, str]] = None,
-        **kwargs: Any,
+            self,
+            embedding: List[float],
+            k: int = 4,
+            fetch_k: int = 20,
+            lambda_mult: float = 0.5,
+            filter: Optional[Dict[str, str]] = None,
+            **kwargs: Any,
     ) -> List[Document]:
-        """Return docs selected using the maximal marginal relevance."""
+        """Return docs selected using the maximal marginal relevance
+            to embedding vector.
 
-        # This is a temporary workaround to make the similarity search
-        # asynchronous. The proper solution is to make the similarity search
-        # asynchronous in the vector store implementations.
-        return await run_in_executor(
-            None,
-            self.max_marginal_relevance_search_by_vector,
+        Maximal marginal relevance optimizes for similarity to query AND diversity
+            among selected documents.
+
+        Args:
+            embedding (str): Text to look up documents similar to.
+            k (int): Number of Documents to return. Defaults to 4.
+            fetch_k (int): Number of Documents to fetch to pass to MMR algorithm.
+                Defaults to 20.
+            lambda_mult (float): Number between 0 and 1 that determines the degree
+                of diversity among the results with 0 corresponding
+                to maximum diversity and 1 to minimum diversity.
+                Defaults to 0.5.
+            filter (Optional[Dict[str, str]]): Filter by metadata. Defaults to None.
+
+        Returns:
+            List[Document]: List of Documents selected by maximal marginal relevance.
+        """
+        docs_and_scores = await self.amax_marginal_relevance_search_with_score_by_vector(
             embedding,
             k=k,
             fetch_k=fetch_k,
@@ -1358,3 +2098,5 @@ class PGVector(VectorStore):
             filter=filter,
             **kwargs,
         )
+
+        return _results_to_docs(docs_and_scores)
