+++ tests/unit_tests/test_vectorstore.py
@@ -5,10 +5,12 @@ from typing import Any, Dict, Generator, List
 import pytest
 from langchain_core.documents import Document
 
-from langchain_postgres.vectorstores_ppr import (
+from langchain_postgres.vectorstores import (
     SUPPORTED_OPERATORS,
     PGVector,
 )
+from sqlalchemy import select
+
 from tests.unit_tests.fake_embeddings import FakeEmbeddings
 from tests.unit_tests.fixtures.filtering_test_cases import (
     DOCUMENTS,
@@ -38,7 +40,7 @@ class FakeEmbeddingsWithAdaDimension(FakeEmbeddings):
         return [float(1.0)] * (ADA_TOKEN_COUNT - 1) + [float(0.0)]
 
 
-def test_pgvector(pgvector: PGVector) -> None:
+def test_pgvector() -> None:
     """Test end to end construction and search."""
     texts = ["foo", "bar", "baz"]
     docsearch = PGVector.from_texts(
@@ -49,6 +51,21 @@ def test_pgvector(pgvector: PGVector) -> None:
         pre_delete_collection=True,
     )
     output = docsearch.similarity_search("foo", k=1)
+    assert output == [Document(page_content="foo")]
+
+# @pytest.mark.requires("xxx") # FIXME
+@pytest.mark.asyncio
+async def test_async_pgvector() -> None:
+    """Test end to end construction and search."""
+    texts = ["foo", "bar", "baz"]
+    docsearch = await PGVector.afrom_texts(
+        texts=texts,
+        collection_name="test_collection",
+        embedding=FakeEmbeddingsWithAdaDimension(),
+        connection=CONNECTION_STRING,
+        pre_delete_collection=True,
+    )
+    output = await docsearch.asimilarity_search("foo", k=1)
     assert output == [Document(page_content="foo")]
 
 
@@ -68,6 +85,23 @@ def test_pgvector_embeddings() -> None:
     assert output == [Document(page_content="foo")]
 
 
+@pytest.mark.asyncio
+async def test_async_pgvector_embeddings() -> None:
+    """Test end to end construction with embeddings and search."""
+    texts = ["foo", "bar", "baz"]
+    text_embeddings = FakeEmbeddingsWithAdaDimension().embed_documents(texts)
+    text_embedding_pairs = list(zip(texts, text_embeddings))
+    docsearch = await PGVector.afrom_embeddings(
+        text_embeddings=text_embedding_pairs,
+        collection_name="test_collection",
+        embedding=FakeEmbeddingsWithAdaDimension(),
+        connection=CONNECTION_STRING,
+        pre_delete_collection=True,
+    )
+    output = await docsearch.asimilarity_search("foo", k=1)
+    assert output == [Document(page_content="foo")]
+
+
 def test_pgvector_with_metadatas() -> None:
     """Test end to end construction and search."""
     texts = ["foo", "bar", "baz"]
@@ -81,6 +115,23 @@ def test_pgvector_with_metadatas() -> None:
         pre_delete_collection=True,
     )
     output = docsearch.similarity_search("foo", k=1)
+    assert output == [Document(page_content="foo", metadata={"page": "0"})]
+
+
+@pytest.mark.asyncio
+async def test_async_pgvector_with_metadatas() -> None:
+    """Test end to end construction and search."""
+    texts = ["foo", "bar", "baz"]
+    metadatas = [{"page": str(i)} for i in range(len(texts))]
+    docsearch = await PGVector.afrom_texts(
+        texts=texts,
+        collection_name="test_collection",
+        embedding=FakeEmbeddingsWithAdaDimension(),
+        metadatas=metadatas,
+        connection=CONNECTION_STRING,
+        pre_delete_collection=True,
+    )
+    output = await docsearch.asimilarity_search("foo", k=1)
     assert output == [Document(page_content="foo", metadata={"page": "0"})]
 
 
@@ -100,6 +151,23 @@ def test_pgvector_with_metadatas_with_scores() -> None:
     assert output == [(Document(page_content="foo", metadata={"page": "0"}), 0.0)]
 
 
+@pytest.mark.asyncio
+async def test_async_pgvector_with_metadatas_with_scores() -> None:
+    """Test end to end construction and search."""
+    texts = ["foo", "bar", "baz"]
+    metadatas = [{"page": str(i)} for i in range(len(texts))]
+    docsearch = await PGVector.afrom_texts(
+        texts=texts,
+        collection_name="test_collection",
+        embedding=FakeEmbeddingsWithAdaDimension(),
+        metadatas=metadatas,
+        connection=CONNECTION_STRING,
+        pre_delete_collection=True,
+    )
+    output = await docsearch.asimilarity_search_with_score("foo", k=1)
+    assert output == [(Document(page_content="foo", metadata={"page": "0"}), 0.0)]
+
+
 def test_pgvector_with_filter_match() -> None:
     """Test end to end construction and search."""
     texts = ["foo", "bar", "baz"]
@@ -113,6 +181,23 @@ def test_pgvector_with_filter_match() -> None:
         pre_delete_collection=True,
     )
     output = docsearch.similarity_search_with_score("foo", k=1, filter={"page": "0"})
+    assert output == [(Document(page_content="foo", metadata={"page": "0"}), 0.0)]
+
+
+@pytest.mark.asyncio
+async def test_async_pgvector_with_filter_match() -> None:
+    """Test end to end construction and search."""
+    texts = ["foo", "bar", "baz"]
+    metadatas = [{"page": str(i)} for i in range(len(texts))]
+    docsearch = await PGVector.afrom_texts(
+        texts=texts,
+        collection_name="test_collection_filter",
+        embedding=FakeEmbeddingsWithAdaDimension(),
+        metadatas=metadatas,
+        connection=CONNECTION_STRING,
+        pre_delete_collection=True,
+    )
+    output = await docsearch.asimilarity_search_with_score("foo", k=1, filter={"page": "0"})
     assert output == [(Document(page_content="foo", metadata={"page": "0"}), 0.0)]
 
 
@@ -133,6 +218,24 @@ def test_pgvector_with_filter_distant_match() -> None:
         (Document(page_content="baz", metadata={"page": "2"}), 0.0013003906671379406)
     ]
 
+@pytest.mark.asyncio
+async def test_async_pgvector_with_filter_distant_match() -> None:
+    """Test end to end construction and search."""
+    texts = ["foo", "bar", "baz"]
+    metadatas = [{"page": str(i)} for i in range(len(texts))]
+    docsearch = await PGVector.afrom_texts(
+        texts=texts,
+        collection_name="test_collection_filter",
+        embedding=FakeEmbeddingsWithAdaDimension(),
+        metadatas=metadatas,
+        connection=CONNECTION_STRING,
+        pre_delete_collection=True,
+    )
+    output = await docsearch.asimilarity_search_with_score("foo", k=1, filter={"page": "2"})
+    assert output == [
+        (Document(page_content="baz", metadata={"page": "2"}), 0.0013003906671379406)
+    ]
+
 
 def test_pgvector_with_filter_no_match() -> None:
     """Test end to end construction and search."""
@@ -147,6 +250,22 @@ def test_pgvector_with_filter_no_match() -> None:
         pre_delete_collection=True,
     )
     output = docsearch.similarity_search_with_score("foo", k=1, filter={"page": "5"})
+    assert output == []
+
+@pytest.mark.asyncio
+async def test_async_pgvector_with_filter_no_match() -> None:
+    """Test end to end construction and search."""
+    texts = ["foo", "bar", "baz"]
+    metadatas = [{"page": str(i)} for i in range(len(texts))]
+    docsearch = await PGVector.afrom_texts(
+        texts=texts,
+        collection_name="test_collection_filter",
+        embedding=FakeEmbeddingsWithAdaDimension(),
+        metadatas=metadatas,
+        connection=CONNECTION_STRING,
+        pre_delete_collection=True,
+    )
+    output = await docsearch.asimilarity_search_with_score("foo", k=1, filter={"page": "5"})
     assert output == []
 
 
@@ -166,6 +285,26 @@ def test_pgvector_collection_with_metadata() -> None:
         else:
             assert collection.name == "test_collection"
             assert collection.cmetadata == {"foo": "bar"}
+
+@pytest.mark.asyncio
+async def test_async_pgvector_collection_with_metadata() -> None:
+    """Test end to end collection construction"""
+    pgvector = await PGVector.create(
+        collection_name="test_collection",
+        collection_metadata={"foo": "bar"},
+        embeddings=FakeEmbeddingsWithAdaDimension(),
+        connection=CONNECTION_STRING,
+        pre_delete_collection=True,
+    )
+    async with pgvector._session_maker() as session:
+        collection = await pgvector.aget_collection(session)
+        if collection is None:
+            assert False, "Expected a CollectionStore object but received None"
+        else:
+            assert collection.name == "test_collection"
+            assert collection.cmetadata == {"foo": "bar"}
+
+
 
 
 def test_pgvector_delete_docs() -> None:
@@ -191,6 +330,35 @@ def test_pgvector_delete_docs() -> None:
     vectorstore.delete(["2", "3"])  # Should not raise on missing ids
     with vectorstore._session_maker() as session:
         records = list(session.query(vectorstore.EmbeddingStore).all())
+        # ignoring type error since mypy cannot determine whether
+        # the list is sortable
+        assert sorted(record.id for record in records) == []  # type: ignore
+
+
+@pytest.mark.asyncio
+async def test_async_pgvector_delete_docs() -> None:
+    """Add and delete documents."""
+    texts = ["foo", "bar", "baz"]
+    metadatas = [{"page": str(i)} for i in range(len(texts))]
+    vectorstore = await PGVector.afrom_texts(
+        texts=texts,
+        collection_name="test_collection_filter",
+        embedding=FakeEmbeddingsWithAdaDimension(),
+        metadatas=metadatas,
+        ids=["1", "2", "3"],
+        connection=CONNECTION_STRING,
+        pre_delete_collection=True,
+    )
+    await vectorstore.adelete(["1", "2"])
+    async with vectorstore._session_maker() as session:
+        records=(await session.execute(select(vectorstore.EmbeddingStore))).scalars().all()
+        # ignoring type error since mypy cannot determine whether
+        # the list is sortable
+        assert sorted(record.id for record in records) == ["3"]  # type: ignore
+
+    await vectorstore.adelete(["2", "3"])  # Should not raise on missing ids
+    async with vectorstore._session_maker() as session:
+        records=(await session.execute(select(vectorstore.EmbeddingStore))).scalars().all()
         # ignoring type error since mypy cannot determine whether
         # the list is sortable
         assert sorted(record.id for record in records) == []  # type: ignore
@@ -485,17 +653,6 @@ def test_pgvector_with_with_metadata_filters_5(
     assert [doc.metadata["id"] for doc in docs] == expected_ids, test_filter
 
 
-@pytest.mark.parametrize("test_filter, expected_ids", TYPE_6_FILTERING_TEST_CASES)
-def test_pgvector_with_with_metadata_filters_6(
-    pgvector: PGVector,
-    test_filter: Dict[str, Any],
-    expected_ids: List[int],
-) -> None:
-    """Test end to end construction and search."""
-    docs = pgvector.similarity_search("meow", k=5, filter=test_filter)
-    assert [doc.metadata["id"] for doc in docs] == expected_ids, test_filter
-
-
 @pytest.mark.parametrize(
     "invalid_filter",
     [
@@ -508,8 +665,6 @@ def test_pgvector_with_with_metadata_filters_6(
         {"$and": {}},
         {"$between": {}},
         {"$eq": {}},
-        {"$exists": {}},
-        {"$exists": 1},
     ],
 )
 def test_invalid_filters(pgvector: PGVector, invalid_filter: Any) -> None:
@@ -524,7 +679,6 @@ def test_validate_operators() -> None:
         "$and",
         "$between",
         "$eq",
-        "$exists",
         "$gt",
         "$gte",
         "$ilike",
